models/base_model.py
#!/usr/bin/python3
"""
BaseModel - Module
BaseModel Parent class
"""

import uuid
from datetime import datetime


class BaseModel:
    def __init__(self, *args, **kwargs):
        if not kwargs:
            self.id = str(uuid.uuid4())
            self.created_at = datetime.now()
            self.updated_at = datetime.now()
            storage.new(self)
        else:
            for key, value in kwargs.items():
                if key != "__class__":
                    if key in ["created_at", "updated_at"]:
                        setattr(self, key, datetime.strptime(value, "%Y-%m-%dT%H:%M:%S.%f"))
                    else:
                        setattr(self, key, value)

    def __str__(self):
        return "[{}] ({}) {}".format(self.__class__.__name, self.id, self.__dict)

    def __repr__(self):
        return "[{}] ({}) {}".format(self.__class__.__name, self.id, self.__dict)

    def save(self):
        self.updated_at = datetime.now()
        storage.save()

    def to_dict(self):
        obj_dict = dict(self.__dict)
        obj_dict['__class__'] = self.__class__.__name
        obj_dict['created_at'] = self.created_at.isoformat()
        obj_dict['updated_at'] = self.updated_at.isoformat()
        return obj_dict
